+++ NEWS.md
@@ -1,3 +1,15 @@
+# RcppThread 2.1.0
+
+NEW FEATURE
+
+* safe printing to the R error stream with `RcppThread::Rcerr` (#60, thanks to
+  @appelmar).
+
+BUG FIX
+
+* adapt thread affinity to possibly restricted CPU set (#61).
+
+
 # RcppThread 2.0.1
 
 BUG FIX